move employee verification, add context prop
